Updated Store section

+++ Steam.Application/DTOs/Store/Coupon/CouponCreateDto.cs
@@ -9,7 +9,7 @@ namespace Steam.Application.DTOs.Store.Coupon
     public record CouponCreateDto
     {
         public string Code { get; init; } = default!;
-        public double Percentage { get; init; }
+        public decimal Percentage { get; init; }
         public DateTime ExpirationDate { get; init; }
     }
 }
